Use GitHub README for plugins.jenkins.io docs
https://www.jenkins.io/doc/developer/publishing/documentation/
describes the process to change from wiki based plugin documentation
to host the plugin documentation in GitHub.

In this case, the preferred plugin documentation location seems to be
the bitbar.com site.  Will adjust the README to first include a link
to the bitbar.com site while retaining the other information in the
README.