Require Jenkins 2.361.4 or newer
The plugin already requires Java 11. Jenkins 2.361.1 is
the first long term support release that requires Java 11.
https://www.jenkins.io/blog/2022/06/28/require-java-11/ announced that
change in June 2022.

https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/
recommends choosing the terminal release (2.361.4) rather than choosing the
initial release in an LTS baseline, so 2.361.4 was chosen instead of 2.361.1.

https://stats.jenkins.io/pluginversions/testdroid-run-in-cloud.html shows
that users of this plugin running Jenkins versions that require Java 11
(2.357 weekly or newer, 2.361.1 LTS or newer) are already running Jenkins
2.375.2 or newer.  This change does not create any barrier for their
upgrade to a release based on this change.  They are already running a
newer version than the minimum Jenkins version required by this change.

Declares dependency on several Jenkins API plugins for testdroid-api
rather than relying on Maven to bundle the transitive dependencies.
This reduces the size of the plugin hpi file and allows Jenkins users to
run with the more recent API version rather than requiring a new release
of this plugin.

Earlier changes in this pull request removed the bundling of guava since
a modern guava version is provided with Jenkins core.  Those dependency
changes have reduced the plugin hpi file from over 10 MB to less than
5 MB.